context_propagation/python_application/requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
opentelemetry-api>=1.28.0
opentelemetry-sdk>=1.28.1
openinference-semantic-conventions>=0.1.5
arize-otel==0.5.3
arize-phoenix-otel>=0.1.0 